+++ src/Modules/Tours/Explorer.Tours.Infrastructure/Database/Repositories/CheckpointDatabaseRepository.cs
@@ -19,7 +19,7 @@ namespace Explorer.Tours.Infrastructure.Database.Repositories
         {
             _dbContext = dbContext;
         }
-        public List<Checkpoint> GetByTourId(int tourId)
+        public List<Checkpoint> GetByTourId(long tourId)
         {
                 return _dbContext.Checkpoints
                     .Where(p => p.TourId == tourId)
